feat：yaml 方式下，mr 触发器支持配置是否阻塞 mr 合并 #9412
