+++ src/backend/ci/core/stream/biz-stream/src/main/kotlin/com/tencent/devops/stream/trigger/listener/StreamBuildFinishListenerService.kt
@@ -27,23 +27,18 @@
 
 package com.tencent.devops.stream.trigger.listener
 
-import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
 import com.tencent.devops.common.api.exception.OperationException
-import com.tencent.devops.common.api.util.YamlUtil
 import com.tencent.devops.common.event.pojo.pipeline.PipelineBuildFinishBroadCastEvent
 import com.tencent.devops.common.pipeline.enums.BuildStatus
 import com.tencent.devops.process.yaml.v2.enums.StreamObjectKind
-import com.tencent.devops.process.yaml.v2.parsers.template.models.NoReplaceTemplate
-import com.tencent.devops.process.yaml.v2.utils.ScriptYmlUtils
 import com.tencent.devops.stream.config.StreamGitConfig
 import com.tencent.devops.stream.dao.GitPipelineResourceDao
 import com.tencent.devops.stream.dao.GitRequestEventBuildDao
 import com.tencent.devops.stream.dao.GitRequestEventDao
 import com.tencent.devops.stream.dao.StreamBasicSettingDao
 import com.tencent.devops.stream.pojo.StreamCIInfo
-import com.tencent.devops.stream.trigger.actions.BaseAction
 import com.tencent.devops.stream.trigger.actions.EventActionFactory
 import com.tencent.devops.stream.trigger.actions.data.StreamTriggerPipeline
 import com.tencent.devops.stream.trigger.actions.data.StreamTriggerSetting
@@ -166,7 +161,6 @@ class StreamBuildFinishListenerService @Autowired constructor(
         )
         action.data.context.requestEventId = requestEvent.id
         action.data.context.extensionAction = requestEvent.extensionAction
-        fixSettingFromYaml(action)
 
         // 推送结束构建消息
         sendCommitCheck.sendCommitCheck(action)
@@ -188,20 +182,5 @@ class StreamBuildFinishListenerService @Autowired constructor(
 
         // 发送通知
         sendNotify.sendNotify(action)
-    }
-
-    private fun fixSettingFromYaml(action: BaseAction) {
-        val realYaml = ScriptYmlUtils.formatYaml(action.data.context.originYaml!!)
-        val preTriggerOn = YamlUtil.getObjectMapper().readValue(
-            realYaml, object : TypeReference<NoReplaceTemplate>() {}
-        ).triggerOn
-        val triggerOn = ScriptYmlUtils.formatTriggerOn(preTriggerOn)
-        if (triggerOn.mr?.reportCommitCheck != null) {
-            action.data.setting.enableCommitCheck = triggerOn.mr?.reportCommitCheck!!
-        }
-
-        if (triggerOn.mr?.blockMr != null) {
-            action.data.setting.enableMrBlock = triggerOn.mr?.blockMr!!
-        }
     }
 }
